+++ app/build.gradle
@@ -26,6 +26,17 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+
+    flavorDimensions "serverType"
+    productFlavors {
+        production {
+            dimension "serverType"
+            buildConfigField "String", "IMAGE_URL", "\"https://image.tmdb.org/t/p/w500/\""
+            buildConfigField "String", "BASE_URL", "\"https://api.themoviedb.org/3/\""
+            buildConfigField "String", "API_KEY", "\"c589a9f6e9e9d8b8b1a3612f1b750053\""
+        }
+    }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
