+++ app/src/main/java/an/maguste/android/navier/moviesdetail/MoviesDetailViewModelFactory.kt
@@ -1,10 +1,9 @@
 package an.maguste.android.navier.moviesdetail
 
-import an.maguste.android.navier.data.Movie
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
-class MoviesDetailViewModelFactory() : ViewModelProvider.Factory{
+class MoviesDetailViewModelFactory : ViewModelProvider.Factory{
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel?> create(modelClass: Class<T>): T = when (modelClass) {
