+++ app/src/main/java/an/maguste/android/navier/moviesdetail/MoviesDetailsViewModel.kt
@@ -3,7 +3,6 @@ package an.maguste.android.navier.moviesdetail
 import an.maguste.android.navier.api.MovieDbApiService
 import an.maguste.android.navier.data.Movie
 import an.maguste.android.navier.data.State
-import an.maguste.android.navier.data.loadMovies
 import an.maguste.android.navier.movieslist.MoviesListViewModel
 import android.util.Log
 import androidx.lifecycle.LiveData
@@ -14,19 +13,13 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import java.lang.Exception
 
-class MoviesDetailsViewModel() : ViewModel() {
+class MoviesDetailsViewModel : ViewModel() {
 
     private val _selectedMovie = MutableLiveData<Movie>()
     val selectedMovie: LiveData<Movie> get() = _selectedMovie
 
     private val _state = MutableLiveData<State>(State.Init)
     val state: LiveData<State> get() = _state
-/*
-    fun setMovie(){
-        _selectedMovie.value = movie
-    }
-*/
-
 
     fun loadMovie(movieId: Int){
         viewModelScope.launch {
@@ -43,16 +36,5 @@ class MoviesDetailsViewModel() : ViewModel() {
                 Log.e(MoviesListViewModel::class.java.simpleName,"Error grab movies data ${e.message}")
             }
         }
-
-   /*     viewModelScope.launch {
-            Log.d("RetrofitTry", "try to load genres")
-
-            //val resultRequest = movieApi.getGenresList()
-            val resultRequest = MovieDbApiService.retrofitService.getMovie(movieId = movieId)
-            Log.d("RetrofitTry", "finish: $resultRequest")
-        }*/
-
     }
-
-
 }
