+++ app/src/main/java/an/maguste/android/navier/movieslist/MoviesListViewModelFactory.kt
@@ -1,13 +1,10 @@
 package an.maguste.android.navier.movieslist
 
 import an.maguste.android.navier.App
-import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
 
-class MoviesListViewModelFactory() : ViewModelProvider.Factory{
+class MoviesListViewModelFactory : ViewModelProvider.Factory{
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel?> create(modelClass: Class<T>): T = when (modelClass) {
