+++ app/src/main/res/layout/fragment_movies_details.xml
@@ -69,12 +69,6 @@
             app:layout_constraintTop_toBottomOf="@+id/poster"
             tools:text="Avengers: \nEnd Game" />
 
-        <!--text shadow-->
-        <!--android:shadowColor="@color/accent"
-        android:shadowDx="2"
-        android:shadowDy="5"
-        android:shadowRadius="5"-->
-
         <TextView
             android:id="@+id/genres"
             style="@style/TextStyle"
@@ -100,9 +94,11 @@
             android:isIndicator="true"
             android:numStars="5"
             android:theme="@style/RatingBarStyle"
+            android:visibility="invisible"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/genres"
-            tools:rating="4" />
+            tools:rating="4"
+            tools:visibility="visible"/>
 
         <TextView
             android:id="@+id/reviews"
@@ -118,18 +114,20 @@
             tools:text="@string/movie_reviews" />
 
         <TextView
-            android:id="@+id/tvStoryline"
+            android:id="@+id/storylineLabel"
             style="@style/TextHeaderStyle"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="@dimen/layout_gap"
             android:layout_marginTop="@dimen/view_big_gap"
             android:text="@string/storyline"
+            android:visibility="invisible"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/reviews" />
+            app:layout_constraintTop_toBottomOf="@+id/reviews"
+            tools:visibility="visible"/>
 
         <TextView
-            android:id="@+id/storyline_text"
+            android:id="@+id/storylineText"
             style="@style/TextRegularStyle"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
@@ -140,7 +138,7 @@
             android:paddingBottom="@dimen/view_medium_gap"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/tvStoryline"
+            app:layout_constraintTop_toBottomOf="@+id/storylineLabel"
             tools:text="@string/movie_description" />
 
         <TextView
@@ -151,8 +149,10 @@
             android:layout_marginStart="@dimen/layout_gap"
             android:layout_marginTop="@dimen/view_big_gap"
             android:text="@string/cast"
+            android:visibility="invisible"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/storyline_text" />
+            app:layout_constraintTop_toBottomOf="@+id/storylineText"
+            tools:visibility="visible"/>
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView"
